Account - READ COMMENT
Right. For the account to work right now you have to manually input the
price you bought shares for, and the amount you made when you sell
shares, manually. This is because we aren't getting the messages from
the server through the interface, they're just being printed out in
receivemessages. If you want the automation you'll have to change
everything so we can get the message in interface and to get the price
and profit automatically.